refactor: move functions to utils to help reusabilty
