+++ frontend/bacmet/app/utils.ts
@@ -1,3 +1,6 @@
+
+import { useState, useEffect, useRef } from "react";
+
 export function navigateInPage(params: Record<string, string | string[]> | URLSearchParams) {
     if (params instanceof URLSearchParams) {
         window.history.pushState(null, "", `?${params.toString()}`);
@@ -6,4 +9,30 @@ export function navigateInPage(params: Record<string, string | string[]> | URLSe
         const searchParams = new URLSearchParams(init);
         window.history.pushState(null, "", `?${searchParams.toString()}`);
     }
+}
+
+export function usePromiseData<T, D>(promiseGenerator: () => Promise<T>, defaultValue: D): T | D {
+  const [data, setData] = useState<T | null>(null);
+  const ref = useRef<() => Promise<T> | null>(null);
+
+  useEffect(() => {
+    ref.current = promiseGenerator;
+    const fetchData = async () => {
+      const result = await promiseGenerator();
+      if (ref.current === promiseGenerator) {
+        setData(result);
+      }
+    }
+    fetchData();
+    return () => {
+      ref.current = null;
+    };
+  }, [setData, promiseGenerator]);
+
+  return data || defaultValue;
+}
+
+export async function fetchData<T>(url: string): Promise<T> {
+  const response = await fetch(url);
+  return await response.json() as T
 }
